Rename helper method
