+++ packages/domain-search/src/helpers/get-featured-suggestions.ts
